+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * Add missing params to `_getImageOptions`
+* Support for images not 1:1 in configurator, where the resize operation take into account the width and height instead of size - [ripe-white/#943](https://github.com/ripe-tech/ripe-white/issues/943)
 
 ### Changed